+++ packages/client/test/shared/index.ts
@@ -1,1 +1,2 @@
 export * from "./values";
+export * from "./session";
